+++ app/src/main/java/com/icecrown/onyxridgecm/enums/FilterType.java
@@ -1,0 +1,19 @@
+//**************************************************************
+// Project: OnyxRidge Construction Management
+//
+// File: FilterType.java
+//
+// Written by: Raymond O'Neill
+//
+// Written on: 12/1/2020
+//
+// Purpose: Used to contain enumerations for filtering. Since
+//          as of 12/1/2020, `Photo` has no field to logically
+//          filter, FilterType is not used in `BrowsePhotos...`
+//*************************************************************
+package com.icecrown.onyxridgecm.enums;
+
+public enum FilterType {
+    NO_FILTER,
+    ACCIDENT_HAPPENED
+}