+++ app/src/main/java/com/icecrown/onyxridgecm/enums/SortType.java
@@ -1,0 +1,21 @@
+//**************************************************************
+// Project: OnyxRidge Construction Management
+//
+// File: SortType.java
+//
+// Written by: Raymond O'Neill
+//
+// Written on: 12/1/2020
+//
+// Purpose: Used to contain enumerations for sorting. Used by
+//          both the `BrowsePhotos...java` and `BrowseReports...java`
+//          to sort the content by a given attribute
+//*************************************************************
+package com.icecrown.onyxridgecm.enums;
+
+public enum SortType {
+    LAST_NAME_ASC,
+    LAST_NAME_DESC,
+    DATE_ASC,
+    DATE_DESC
+}