+++ app/src/main/java/com/icecrown/onyxridgecm/fragments/BrowsePhotosFragment.java
@@ -28,6 +28,13 @@
 //      - Implemented the use of `LocalDate` and removed `Date`
 //        instances.
 //      - Removed unused imports
+// ------------------------------------------------
+// - 12/1/2020
+// - R.O.
+// - DETAILS:
+//      - Added code to handle filtering and sorting
+//        - Since there isn't any filtering type for photos
+//        (yet), filtering for photos is currently disabled.
 //**************************************************************
 package com.icecrown.onyxridgecm.fragments;
 
@@ -37,6 +44,7 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.PopupMenu;
 
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
@@ -52,6 +60,7 @@ import com.google.firebase.storage.StorageMetadata;
 import com.google.firebase.storage.StorageReference;
 import com.icecrown.onyxridgecm.R;
 import com.icecrown.onyxridgecm.adapters.PhotosAdapter;
+import com.icecrown.onyxridgecm.enums.SortType;
 import com.icecrown.onyxridgecm.interfaces.IProjectSelectedCallback;
 import com.icecrown.onyxridgecm.utility.Photo;
 import com.icecrown.onyxridgecm.utility.PhotoFactory;
@@ -61,6 +70,7 @@ import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 
 public class BrowsePhotosFragment extends Fragment {
@@ -72,7 +82,8 @@ public class BrowsePhotosFragment extends Fragment {
     private PhotosAdapter photosAdapter;
     private final List<Photo> photoList = new ArrayList<>();
     private FragmentManager manager;
-    private LocalDate dateOfPhoto;
+    private MaterialTextView sortTV;
+    private MaterialTextView filterTV;
 
 
 
@@ -90,6 +101,55 @@ public class BrowsePhotosFragment extends Fragment {
         manager = getActivity().getSupportFragmentManager();
 
         View v = inflater.inflate(R.layout.fragment_browse_content, container, false);
+
+        sortTV = v.findViewById(R.id.sort_text_view);
+        sortTV.setOnClickListener(l -> {
+            PopupMenu sortOptions = new PopupMenu(getActivity(), sortTV);
+            sortOptions.getMenuInflater().inflate(R.menu.menu_sort_options, sortOptions.getMenu());
+            sortOptions.setOnMenuItemClickListener(item -> {
+                int itemId = item.getItemId();
+                if(itemId == R.id.no_sort) {
+                    Log.d("EPOCH-3", "No sort entered.");
+                    photosAdapter.setPhotos(photoList);
+                }
+                else if(itemId == R.id.last_name_sort_asc) {
+                    Log.d("EPOCH-3", "last name asc sort entered.");
+                    sortPhotosByOption(photosAdapter.getPhotos(), SortType.LAST_NAME_ASC);
+                }
+                else if(itemId == R.id.last_name_sort_desc) {
+                    Log.d("EPOCH-3", "last name desc sort entered.");
+                    sortPhotosByOption(photosAdapter.getPhotos(), SortType.LAST_NAME_DESC);
+                }
+                else if(itemId == R.id.date_sort_asc) {
+                    Log.d("EPOCH-3", "date asc sort entered.");
+                    sortPhotosByOption(photosAdapter.getPhotos(), SortType.DATE_ASC);
+                }
+                else if(itemId == R.id.date_sort_desc) {
+                    Log.d("EPOCH-3", "date desc sort entered.");
+                    sortPhotosByOption(photosAdapter.getPhotos(), SortType.DATE_DESC);
+                }
+                photosAdapter.notifyDataSetChanged();
+                return true;
+            });
+            sortOptions.show();
+        });
+        filterTV = v.findViewById(R.id.filter_text_view);
+        filterTV.setOnClickListener(l -> {
+            PopupMenu filterOptions = new PopupMenu(getActivity(), filterTV);
+            filterOptions.getMenuInflater().inflate(R.menu.menu_filter_options, filterOptions.getMenu());
+            filterOptions.setOnMenuItemClickListener(item -> {
+                int itemId = item.getItemId();
+                if(itemId == R.id.no_filter) {
+                    photosAdapter.setPhotos(photoList);
+                }
+                else if(itemId == R.id.accident_filter) {
+                    Snackbar.make(filterTV, R.string.filter_option_not_valid, Snackbar.LENGTH_SHORT).show();
+                }
+                photosAdapter.notifyDataSetChanged();
+                return true;
+            });
+            filterOptions.show();
+        });
 
         projectNameTextView = v.findViewById(R.id.project_name);
 
@@ -142,11 +202,20 @@ public class BrowsePhotosFragment extends Fragment {
                             photo.getMetadata().addOnCompleteListener(task1 -> {
                                 if(task.isSuccessful()) {
                                     final StorageMetadata meta = task1.getResult();
+                                    LocalDate dateOfPhoto = null;
                                     try {
                                         dateOfPhoto = LocalDate.parse(meta.getCustomMetadata("date_uploaded"), DateTimeFormatter.ofPattern("MM/dd/yyyy"));
+                                        Log.d("EPOCH-3", "Date first format successfully parsed.");
                                     } catch (Exception e) {
-                                        Log.d("EPOCH-3", "Date exception for 'date_uploaded' occurred.");
-                                        e.printStackTrace();
+                                        Log.d("EPOCH-3", "DateFormatter try/catch entered");
+                                        try {
+                                            dateOfPhoto = LocalDate.parse(meta.getCustomMetadata("date_uploaded"), DateTimeFormatter.ofPattern("MM/d/yyyy"));
+                                            Log.d("EPOCH-3", "Date second format successfully parsed.");
+                                        }
+                                        catch (Exception e1) {
+                                            Log.d("EPOCH-3", "Date exception for 'date_uploaded' occurred.");
+                                            e1.printStackTrace();
+                                        }
                                     }
 
                                     File f = null;
@@ -162,8 +231,13 @@ public class BrowsePhotosFragment extends Fragment {
 
                                     final File fileTemp = f;
 
+                                    Log.d("EPOCH-3", "Date format for image: " + dateOfPhoto.toString());
+
+
+                                    final LocalDate finalDateOfPhoto = dateOfPhoto;
+
                                     photo.getFile(fileTemp).addOnCompleteListener(task2 -> {
-                                        Photo adder = new Photo(fileTemp.getName(), Uri.fromFile(fileTemp), photo, dateOfPhoto, meta.getCustomMetadata("taken_by_last"), meta.getCustomMetadata("taken_by_first"));
+                                        Photo adder = new Photo(fileTemp.getName(), Uri.fromFile(fileTemp), photo, finalDateOfPhoto, meta.getCustomMetadata("taken_by_last"), meta.getCustomMetadata("taken_by_first"));
                                         adder.generateBitmapForUri(getActivity(), getActivity().getContentResolver());
                                         photoList.add(adder);
 
@@ -185,5 +259,56 @@ public class BrowsePhotosFragment extends Fragment {
                 Snackbar.make(recView, R.string.no_photos_admin, Snackbar.LENGTH_SHORT).show();
             }
         });
+    }
+
+    private void sortPhotosByOption(List<Photo> photos, SortType type) {
+        Comparator<Photo> c;
+
+        switch(type) {
+            case LAST_NAME_ASC:
+                c = (o1, o2) -> {
+                    int docLastVal = o1.getTakenByLastNameFirst().compareTo(o2.getTakenByLastNameFirst());
+                    return Integer.compare(docLastVal, 0);
+                };
+                break;
+            case LAST_NAME_DESC:
+                c = (o1, o2) -> {
+                    int docLastVal = o1.getTakenByLastNameFirst().compareTo(o2.getTakenByLastNameFirst());
+                    return Integer.compare(docLastVal, 0);
+                };
+                c = c.reversed();
+                break;
+            case DATE_ASC:
+                c = (o1, o2) -> {
+                    if(o1.getDate().isBefore(o2.getDate())) {
+                        return -1;
+                    }
+                    else if (o2.getDate().isAfter(o2.getDate())) {
+                        return 1;
+                    }
+                    else {
+                        return 0;
+                    }
+                };
+                break;
+            case DATE_DESC:
+                c = (o1, o2) -> {
+                    if(o1.getDate().isBefore(o2.getDate())) {
+                        return -1;
+                    }
+                    else if (o2.getDate().isAfter(o2.getDate())) {
+                        return 1;
+                    }
+                    else {
+                        return 0;
+                    }
+                };
+                c = c.reversed();
+                break;
+            default:
+                c = null;
+                break;
+        }
+        photos.sort(c);
     }
 }